api-showcase/pyproject.toml
[project]
name = "api-showcase"
version = "0.1.0"
description = ""
authors = [
    {name = "DanielLanghann",email = "daniel.langhann@lector.ai"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp (>=3.13.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "markdown (>=3.9,<4.0)",
    "weasyprint (>=66.0,<67.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "certifi (>=2025.10.5,<2026.0.0)"
]

[tool.poetry]
packages = [{include = "api_showcase", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
